Create local GameTimer object and map to DTO.

+++ src/main/java/com/rocketden/main/dto/game/GameTimerDto.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 @Setter
 public class GameTimerDto {
     private LocalDateTime startTime;
-    private Integer duration;
+    private Long duration;
     private LocalDateTime endTime;
     private boolean timeUp;
 }
